requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
groq==0.18.0
# Para ejecutar
#pip install -r requirements.txt
#py -m venv venv
#venv\Scripts\activate
#
